Move combines helpers to a new file
